ref(home api): move files from 'recipes' to 'home' folder
