+++ src/router/v1/home/index.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { homeRecipes } from "../../../controllers/recipes/recipesHomeController.js";
+import { homeRecipes } from "../../../controllers/home/homeRecipesController.js";
 
 const router = express.Router();
